share more code
